fix warnings
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@5106 3db68df2-c116-0410-a063-a993310a9797

+++ source/tools/atlas/AtlasUI/ScenarioEditor/Tools/PlaceObject.cpp
@@ -8,7 +8,7 @@
 
 using AtlasMessage::Position;
 
-static float g_DefaultAngle = M_PI*3.0/4.0;
+static float g_DefaultAngle = (float)(M_PI*3.0/4.0);
 
 class PlaceObject : public StateDrivenTool<PlaceObject>
 {
